Update mock generation paths and refactor related tests to use new mock interfaces for Firebase, Mongo, and Redis clients.
