+++ internal/infrastructure/auth/firebase_mock.go
@@ -3,17 +3,16 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=firebase.go -destination=mocks/firebase_mock.go -package=mocks
+//	mockgen -source=firebase.go -destination=firebase_mock.go -package=auth
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package auth is a generated GoMock package.
+package auth
 
 import (
 	context "context"
 	reflect "reflect"
 
-	auth "github.com/Financial-Partner/server/internal/infrastructure/auth"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -42,10 +41,10 @@ func (m *MockFirebaseAuth) EXPECT() *MockFirebaseAuthMockRecorder {
 }
 
 // VerifyIDToken mocks base method.
-func (m *MockFirebaseAuth) VerifyIDToken(ctx context.Context, idToken string) (*auth.Token, error) {
+func (m *MockFirebaseAuth) VerifyIDToken(ctx context.Context, idToken string) (*Token, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "VerifyIDToken", ctx, idToken)
-	ret0, _ := ret[0].(*auth.Token)
+	ret0, _ := ret[0].(*Token)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
