+++ internal/infrastructure/database/mongo_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=mongo.go -destination=../mocks/mongo.go -package=mocks
+//	mockgen -source=mongo.go -destination=mongo_mock.go -package=database
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package database is a generated GoMock package.
+package database
 
 import (
 	context "context"
