+++ internal/infrastructure/persistence/redis/redis_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=redis.go -destination=mocks/redis_mock.go -package=mocks
+//	mockgen -source=redis.go -destination=redis_mock.go -package=redis
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package redis is a generated GoMock package.
+package redis
 
 import (
 	context "context"
