+++ internal/interfaces/http/auth_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=auth.go -destination=mocks/auth_mock.go -package=mocks
+//	mockgen -source=auth.go -destination=auth_mock.go -package=handler
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package handler is a generated GoMock package.
+package handler
 
 import (
 	context "context"
