+++ internal/interfaces/http/goals_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=goals.go -destination=mocks/goals_mock.go -package=mocks
+//	mockgen -source=goals.go -destination=goals_mock.go -package=handler
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package handler is a generated GoMock package.
+package handler
 
 import (
 	context "context"
