+++ internal/interfaces/http/middleware/auth_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=auth.go -destination=mocks/auth_mock.go -package=mocks
+//	mockgen -source=auth.go -destination=auth_mock.go -package=middleware
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package middleware is a generated GoMock package.
+package middleware
 
 import (
 	reflect "reflect"
