+++ internal/interfaces/http/user_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=user.go -destination=mocks/user_mock.go -package=mocks
+//	mockgen -source=user.go -destination=user_mock.go -package=handler
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package handler is a generated GoMock package.
+package handler
 
 import (
 	context "context"
